Criando tabelas, e endpoints de usuários
