+++ semana17/projeto-LabECommerce-BackEnd/src/endpoints/createUser.ts
@@ -1,0 +1,10 @@
+import { Request, Response } from "express";
+import { connection } from "../connection";
+
+export const createUser = async (req: Request, res: Response): Promise<void> => {
+    try {
+        
+    } catch (error) {
+        
+    }
+}
