+++ semana17/projeto-LabECommerce-BackEnd/src/endpoints/getAllUsers.ts
@@ -1,0 +1,10 @@
+import { Request, Response } from "express";
+import { connection } from "../connection";
+
+export const getAllUsers = async (req: Request, res: Response): Promise<void> => {
+    try {
+        
+    } catch (error) {
+        
+    }
+}
