+++ semana17/projeto-LabECommerce-BackEnd/src/index.ts
@@ -1,32 +1,6 @@
-import express, { Express } from 'express'
-import cors from 'cors'
-import { AddressInfo } from "net";
-import knex from "knex";
-import dotenv from "dotenv";
+import app from "./app";
+import { createUser } from "./endpoints/createUser";
+import { getAllUsers } from "./endpoints/getAllUsers";
 
-dotenv.config();
-
-const app: Express = express();
-
-app.use(express.json());
-app.use(cors());
-
-export const connection = knex({
-    client: "mysql",
-    connection: {
-        host: process.env.DB_HOST,
-        port: 3306,
-        user: process.env.DB_USER,
-        password: process.env.DB_PASSWORD,
-        database: process.env.DB_SCHEMA
-    }
-});
-
-const server = app.listen(process.env.PORT || 3003, () => {
-    if (server) {
-        const address = server.address() as AddressInfo;
-        console.log(`Server is running in http://localhost:${address.port}`);
-    } else {
-        console.error(`Failure upon starting server.`);
-    }
-});
+app.get("/users", getAllUsers)
+app.post("/users", createUser)
